Add a function to prevent user to press return
add a function to prevent user to press return when the game is
calculating a win or full board. the button is causing a bug where the
round count is updating its count when user press it in the middle of
the calculation, showing 2 round count on game start instead of 1.